feat: add delete and update customer implementation

+++ QubixCinema/Forms/MovieForms/DeleteMovie.cs
@@ -1,6 +1,6 @@
-﻿using DevExpress.XtraEditors;
+using DevExpress.XtraEditors;
+using QubixCinema.Business.Services;
 using QubixCinema.DataAccess;
-using QubixCinema.DataAccess.Services;
 using QubixCinema.Entities.Models;
 using System;
 using System.Collections.Generic;
@@ -10,7 +10,7 @@ namespace QubixCinema.Forms.MovieForms
 {
     public partial class DeleteMovie : Form
     {
-        private readonly MovieService _movieService;
+        private MovieService _movieService;
         public DeleteMovie()
         {
             InitializeComponent();
@@ -19,7 +19,14 @@ namespace QubixCinema.Forms.MovieForms
 
         private void DeleteMovie_Load(object sender, EventArgs e)
         {
-            LoadMovies();
+            try
+            {
+                LoadMovies();
+            }
+            catch (Exception ex)
+            {
+                System.Windows.Forms.MessageBox.Show($"Load Error: {ex.ToString()}");
+            }
         }
 
         private void LoadMovies()
@@ -38,18 +45,20 @@ namespace QubixCinema.Forms.MovieForms
 
         private void grid_movies_DoubleClick(object sender, EventArgs e)
         {
-            DataGridViewRow chosenRow = grid_movies.CurrentRow;
-            int movie = Convert.ToInt32(chosenRow.Cells[0].Value);
-
-            DialogResult result = XtraMessageBox.Show("Are you sure you want to delete the selected movie?",
-                 "Question", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-
-            if (result == DialogResult.Yes)
+            if (grid_movies.CurrentRow?.DataBoundItem is Movie chosenMovie)
             {
-                _movieService.DeleteMovie(movie);
-                LoadMovies();
-                XtraMessageBox.Show("The movie has been successfully removed",
-                "Successful", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                int movie = chosenMovie.MovieId;
+
+                DialogResult result = XtraMessageBox.Show("Are you sure you want to delete the selected movie?",
+                     "Question", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)
+                {
+                    _movieService.DeleteMovie(movie);
+                    LoadMovies();
+                    XtraMessageBox.Show("The movie has been successfully removed",
+                    "Successful", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
             }
         }
     }
